fix: 시간 연산에서 매개변수 타입에 맞춰 long 캐스팅 적용

+++ src/main/java/flab/transtalk/user/service/user/UserMatchingService.java
@@ -69,7 +69,7 @@ public class UserMatchingService {
         int updatedCount = current + refill;
         status.setRemainingMatchRequests(updatedCount);
 
-        status.setLastMatchRequestedAt(lastRequested.plusHours(refill * ServiceConfigConstants.MATCH_REQUEST_RECHARGE_INTERVAL_HOURS));
+        status.setLastMatchRequestedAt(lastRequested.plusHours((long) refill * ServiceConfigConstants.MATCH_REQUEST_RECHARGE_INTERVAL_HOURS));
     }
 
     @Transactional
